Initial changes suggested by Ben during PR review

+++ src/core/refinement.h
@@ -27,8 +27,8 @@ public:
 	float beam_tilt_y;
 	float image_shift_x;
 	float image_shift_y;
-	float beam_tilt_group; // TODO - why is this not int?
-	float particle_group; // TODO - why is this not int?
+	int beam_tilt_group; 
+	int particle_group; 
 	int assigned_subset; 
 	float pre_exposure;
 	float total_exposure;
